Go to 0.60mm extrusion width with 0.9 multiplier
After deciding it didn't help earlier, later changes dialled other
things in to the point where I got an accurate single wall measurement
of 0.60mm.

This now gives me a 20mm cube measurement of 20mm on the X and Y,
though still a little tall on the Z (20.15mm) with the first layer
height restored to 100%.

Will go back to first layer calibration after this with the new settings.