Recalibrate nozzle based on a 0.5 width
The x1.2 thing seems to be a weird bit of reprap lore, someone who
did testing found the strength was best at the nozzle width.

Calibrating gave me an extrusion multiplier of 0.86, so go with that.

Adjust the top/bottom solid layers to give 1mm, and the walls to be
1mm as well, by default.